Swimlane bar plot for each id with markers

Swimlane bar plot for each id with markers

g_swimlane(bar_id, bar_length, sort_by = NULL, col_by = NULL,
  marker_id = NULL, marker_pos = NULL, marker_shape = NULL,
  marker_shape_opt = NULL, marker_color = NULL, marker_color_opt = NULL,
  anno_txt = NULL, yref_line = NULL, ytick_at = waiver(), ylab, title)

Arguments

bar_id
vector of IDs to identify each bar

bar_length
numeric vector to be plotted as length for each bar

sort_by
vector to sort bars

col_by
vector to color bars

marker_id
vector of IDs to identify markers within each bar. Default is the same as bar_id.

marker_pos
numeric vector to specify position for each marker

marker_shape
vector to specify shape for markers

marker_shape_opt
aesthetic values to map shape values (named vector to map shape values to each name)

marker_color
vector to specify color for markers

marker_color_opt
aesthetic values to map shape values (named vector to map shape values to each name)

anno_txt
dataframe of variables to be displayed as annotation on the left

yref_line
numeric vector to plot reference lines

ytick_at
break interval of y-axis

ylab
y label

title
String to be displayed as plot title

Value

plot object

Examples

library(random.cdisc.data)
library(dplyr)
ASL <- radam("ASL", N=50, start_with = list(TRTDUR = rexp(50, 1/100)))
ARS <- radam("ARS", ADSL = ASL,
start_with = list(
      ADY = rexp(nrow(ASL), 1/80)
   )) %>% filter(PARAMCD == "OVRINV")
ANL <- ASL %>% left_join(ARS, by = c("STUDYID", "USUBJID"))
# pre-process bar length breaks for treatment duration for longitudinal records
anno_txt <- ASL[, c("ARMCD", "SEX", "RACE")]

g_swimlane(bar_id = ASL$USUBJID,
bar_length = ASL$TRTDUR,
sort_by = ASL$ARM,
col_by = ASL$ARM,
marker_id = ANL$USUBJID,
marker_pos = ANL$ADY,
marker_shape = ANL$AVALC,
marker_shape_opt = c("CR" = 16, "PR" = 17, "SD" = 18, "PD" = 15, "NE" = 4),
marker_color = NULL,
marker_color_opt = NULL,
anno_txt = anno_txt,
yref_line = c(100, 200),
ytick_at = waiver(),
# xlab = "Patient ID",
ylab = "Time from First Treatment (Day)",
title = "Swimlane Plot")

\donotrun{
library(dplyr)

atx <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/atx.sas7bdat")
asl <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/asl.sas7bdat")
xars <- read.bce("/opt/BIOSTAT/qa/s30103j/libraries/xars.sas7bdat")
# phase 1A data
# pre-process ASl and ARS to fit in ANL
ATX <- atx %>% filter(grepl("1a", APERIDC2))
ASL <- ATX %>% select(USUBJID) %>%
left_join(asl, by = "USUBJID") %>%
filter(SAFFL == "Y")

anno_txt_var <- c("ARMCD", "SEX", "CADX")
anno_txt <- ASL[, anno_txt_var]

ARS <- ASL %>% select(USUBJID) %>%
left_join(xars %>% filter(grepl("1a", APERIDC2)), "USUBJID") %>%
filter(PARAMCD == "OVRINV") %>%
select(USUBJID, ADY, AVALC)

ADS <- ASL %>%
filter(DISCSTUD == "Y" | !is.na(STDSSDT)) %>%
select(USUBJID, STDDRS, STDSDY) %>%
rename(ADY = STDSDY, AVALC = STDDRS)

# combine with ASL to generate a length for each record
ANL <- ASL %>% select(USUBJID, ARMCD, SEX, CADX, TRTDUR) %>%
inner_join(rbind(ARS, ADS), "USUBJID")

g_swimlane(bar_id = ASL$USUBJID,
bar_length = ASL$TRTDUR,
sort_by = ASL$ARMCD,
col_by = ASL$ARMCD,
marker_id = ANL$USUBJID,
marker_pos = ANL$ADY,
marker_shape = ANL$AVALC,
marker_shape_opt <- c("CR" = 16, "PR" = 17, "SD" = 18, "PD" = 15,
"DEATH" = 8, "LOST TO FOLLOW-UP" = 10, "WITHDRAWAL BY SUBJECT" = 14),
marker_color = ANL$AVALC,
marker_color_opt <- c("CR" = "green", "PR" = "blue", "SD" = "yellow", "PD" = "red",
"DEATH" = "black", "LOST TO FOLLOW-UP" = "purple", "WITHDRAWAL BY SUBJECT" = "darkred"),
anno_txt = anno_txt,
yref_line = c(100, 200),
ytick_at = waiver(),
# xlab = "Patient ID",
ylab = "Time from First Treatment (Day)",
title = "Swimlane Plot")

}

#> Error: <text>:29:1: unexpected input
#> 28: 
#> 29: \
#>     ^