Adverse Events by System Organ Class and Preferred Term

t_ae returns adverse events summary table that corresponds to STREAM template AET02

t_ae(class, term, id, col_by, total = "All Patients")

Arguments

class
system organ class variable.

term
preferred term variable.

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing.

total
character string that will be used as a label for a column with pooled total population, default here is "All Patients", if set to NULL then the "All Patients" column is suppressed.

Value

rtable object

Details

this is an equivalent of the STREAM output %stream_t_summary(templates = aet02) (http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet02.html)

Examples

# Simple example
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1, 2), c(3, 7))])
)

ae_lookup <- tribble(
  ~CLASS, ~TERM, ~GRADE,
  "cl A", "trm A_1/2", 1,
  "cl A", "trm A_2/2", 2,
  "cl B", "trm B_1/3", 2,
  "cl B", "trm B_2/3", 3,
  "cl B", "trm B_3/3", 1,
  "cl C", "trm C_1/1", 1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2, 2, 2, 3, 3, 4, 4, 4, 4, 5, 6, 6, 7, 7)]
  ),
  ae_lookup[c(1, 1, 2, 6, 4, 2, 2, 3, 4, 2, 1, 5, 4, 6), ]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ae(
  class = ANL$CLASS,
  term = ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients"
)
tbl

#> ANL$CLASS                                                            ARM A               ARM B            All Patients
#>   ANL$TERM                                                           (N=3)               (N=7)               (N=10)   
#> ----------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event           2 (66.7%)           4 (57.1%)            6 (60%)   
#> 
#> Overall total number of events                                         5                   9                   14     
#> 
#> cl A
#>   Total number of patients with at least one adverse event         1 (33.3%)           3 (42.9%)            4 (40%)   
#>   Total number of events                                               3                   4                   7      
#>   trm A_2/2                                                        1 (33.3%)           2 (28.6%)            3 (30%)   
#>   trm A_1/2                                                        1 (33.3%)           1 (14.3%)            2 (20%)   
#> 
#> cl B
#>   Total number of patients with at least one adverse event         1 (33.3%)           3 (42.9%)            4 (40%)   
#>   Total number of events                                               1                   4                   5      
#>   trm B_2/3                                                        1 (33.3%)           2 (28.6%)            3 (30%)   
#>   trm B_1/3                                                            0               1 (14.3%)            1 (10%)   
#>   trm B_3/3                                                            0               1 (14.3%)            1 (10%)   
#> 
#> cl C
#>   Total number of patients with at least one adverse event         1 (33.3%)           1 (14.3%)            2 (20%)   
#>   Total number of events                                               1                   1                   2      
#>   trm C_1/1                                                        1 (33.3%)           1 (14.3%)            2 (20%)   

# Simple example 2

ADSL <- rADSL %>% select(USUBJID, STUDYID, ARM)
AAE <- rADAE %>% select(USUBJID, STUDYID, ARM, AEBODSYS, AEDECOD)
ANL <- left_join(AAE, ADSL, by = c("USUBJID", "STUDYID", "ARM"))

tbl2 <- t_ae(
  class = ANL$AEBODSYS,
  term = ANL$AEDECOD,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = NULL
)
tbl2

#> Body System or Organ Class                                                   A: Placebo            B: Drug X           C: Combination
#>   Dictionary Derived Term                                                      (N=11)                (N=21)                (N=18)    
#> -------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event                     11 (100%)             21 (100%)             18 (100%)   
#> 
#> Overall total number of events                                                   65                   133                    99      
#> 
#> Investigations
#>   Total number of patients with at least one adverse event                   6 (54.5%)             12 (57.1%)            10 (55.6%)  
#>   Total number of events                                                         8                     16                    20      
#>   Amniocentesis normal                                                           0                     0                  1 (5.6%)   
#>   Anti factor VIII antibody test                                                 0                     0                  1 (5.6%)   
#>   Anti-aquaporin-4 antibody                                                      0                     0                  1 (5.6%)   
#>   Antibody test negative                                                         0                  1 (4.8%)                 0       
#>   Antidepressant drug clearance increased                                        0                     0                  1 (5.6%)   
#>   Aspergillus test negative                                                   1 (9.1%)                 0                     0       
#>   Beta 2 microglobulin decreased                                                 0                     0                  1 (5.6%)   
#>   Biopsy soft tissue                                                          1 (9.1%)                 0                     0       
#>   Blood beta-D-glucan normal                                                     0                     0                  1 (5.6%)   
#>   Blood bicarbonate                                                              0                     0                  1 (5.6%)   
#>   Blood glucagon normal                                                          0                     0                  1 (5.6%)   
#>   Blood oestrogen abnormal                                                       0                     0                  1 (5.6%)   
#>   Blood pressure normal                                                          0                     0                  1 (5.6%)   
#>   Blood selenium                                                                 0                  1 (4.8%)                 0       
#>   Carnitine abnormal                                                             0                  1 (4.8%)                 0       
#>   Cholecystogram intravenous abnormal                                         1 (9.1%)                 0                     0       
#>   Citric acid urine decreased                                                    0                     0                  1 (5.6%)   
#>   Cytokeratin 18                                                                 0                     0                  1 (5.6%)   
#>   Cytokeratin 18 increased                                                       0                     0                  1 (5.6%)   
#>   Dialysis efficacy test                                                         0                  1 (4.8%)                 0       
#>   Drug level                                                                     0                  1 (4.8%)                 0       
#>   Electrocardiogram PQ interval prolonged                                        0                  1 (4.8%)                 0       
#>   Emergency care examination                                                     0                  1 (4.8%)                 0       
#>   Erythroblast count normal                                                      0                     0                  1 (5.6%)   
#>   Flavivirus test negative                                                    1 (9.1%)                 0                     0       
#>   Hepatitis B DNA assay negative                                              1 (9.1%)                 0                     0       
#>   Hepatitis B e antigen negative                                                 0                  1 (4.8%)                 0       
#>   JC polyomavirus test positive                                                  0                  1 (4.8%)                 0       
#>   Karyotype analysis abnormal                                                 1 (9.1%)                 0                     0       
#>   Metamyelocyte percentage increased                                             0                  1 (4.8%)                 0       
#>   Methaemoglobin urine                                                           0                     0                  1 (5.6%)   
#>   Microbiology test abnormal                                                     0                     0                  1 (5.6%)   
#>   Pancreatic polypeptide level increased                                         0                  1 (4.8%)                 0       
#>   Platelet factor 4                                                              0                     0                  1 (5.6%)   
#>   Pregnancy test false positive                                               1 (9.1%)                 0                     0       
#>   Streptococcus test                                                             0                     0                  1 (5.6%)   
#>   Stroke volume                                                                  0                  1 (4.8%)                 0       
#>   Sweat test                                                                  1 (9.1%)                 0                     0       
#>   T-lymphocyte count abnormal                                                    0                  1 (4.8%)                 0       
#>   Urine homocystine present                                                      0                  1 (4.8%)                 0       
#>   Vitamin B12 increased                                                          0                  1 (4.8%)                 0       
#>   Vitamin K increased                                                            0                     0                  1 (5.6%)   
#>   Wedged hepatic venous pressure                                                 0                  1 (4.8%)                 0       
#>   White blood cell count abnormal                                                0                     0                  1 (5.6%)   
#> 
#> Injury, poisoning and procedural complications
#>   Total number of patients with at least one adverse event                   6 (54.5%)             8 (38.1%)             7 (38.9%)   
#>   Total number of events                                                         10                    12                    7       
#>   Acute graft versus host disease in intestine                                   0                  1 (4.8%)                 0       
#>   Administration site macule                                                  1 (9.1%)                 0                     0       
#>   Application site cyst                                                          0                     0                  1 (5.6%)   
#>   Application site plaque                                                     1 (9.1%)                 0                     0       
#>   Carotid artery perforation                                                  1 (9.1%)                 0                     0       
#>   Drug dose omission by device                                                   0                     0                  1 (5.6%)   
#>   Exposure to noise                                                              0                  1 (4.8%)                 0       
#>   Eye muscle entrapment                                                       1 (9.1%)                 0                     0       
#>   Fluorosis dental                                                            1 (9.1%)                 0                     0       
#>   Infusion site joint infection                                               1 (9.1%)                 0                     0       
#>   Infusion site ulcer                                                            0                  1 (4.8%)                 0       
#>   Injection site laceration                                                      0                  1 (4.8%)                 0       
#>   Medical device site phlebitis                                                  0                  1 (4.8%)                 0       
#>   Mitral perforation                                                             0                     0                  1 (5.6%)   
#>   Ocular toxicity                                                                0                  1 (4.8%)                 0       
#>   Off label use                                                                  0                  1 (4.8%)                 0       
#>   Overdose                                                                       0                     0                  1 (5.6%)   
#>   Postmastectomy lymphoedema syndrome                                         1 (9.1%)                 0                     0       
#>   Product communication issue                                                    0                  1 (4.8%)                 0       
#>   Product label confusion                                                        0                  1 (4.8%)                 0       
#>   Prompt effects of radiation                                                    0                  1 (4.8%)                 0       
#>   Refractoriness to platelet transfusion                                         0                  1 (4.8%)                 0       
#>   Shrapnel wound                                                                 0                     0                  1 (5.6%)   
#>   Tongue injury                                                                  0                  1 (4.8%)                 0       
#>   Transfusion microchimerism                                                  1 (9.1%)                 0                     0       
#>   Vaccination site cyst                                                          0                     0                  1 (5.6%)   
#>   Vessel puncture site occlusion                                                 0                     0                  1 (5.6%)   
#>   Wound dehiscence                                                            1 (9.1%)                 0                     0       
#>   Wrong drug administered                                                     1 (9.1%)                 0                     0       
#> 
#> Infections and infestations
#>   Total number of patients with at least one adverse event                   3 (27.3%)             10 (47.6%)            4 (22.2%)   
#>   Total number of events                                                         4                     17                    6       
#>   Anthrax sepsis                                                                 0                  1 (4.8%)              1 (5.6%)   
#>   Abdominal wall abscess                                                         0                  1 (4.8%)                 0       
#>   AIDS retinopathy                                                               0                  1 (4.8%)                 0       
#>   Burkholderia infection                                                         0                  1 (4.8%)                 0       
#>   Carbuncle                                                                      0                  1 (4.8%)                 0       
#>   Chagas' cardiomyopathy                                                         0                  1 (4.8%)                 0       
#>   Cholera                                                                        0                  1 (4.8%)                 0       
#>   Denture stomatitis                                                             0                  1 (4.8%)                 0       
#>   Ear infection fungal                                                           0                     0                  1 (5.6%)   
#>   Epidermodysplasia verruciformis                                                0                     0                  1 (5.6%)   
#>   Gastroenteritis viral                                                       1 (9.1%)                 0                     0       
#>   Hepatitis infectious mononucleosis                                             0                  1 (4.8%)                 0       
#>   Hepatobiliary infection                                                        0                     0                  1 (5.6%)   
#>   Human anaplasmosis                                                          1 (9.1%)                 0                     0       
#>   Infestation                                                                    0                  1 (4.8%)                 0       
#>   JC virus infection                                                             0                  1 (4.8%)                 0       
#>   Lineal gingival erythema                                                       0                  1 (4.8%)                 0       
#>   Miliary pneumonia                                                           1 (9.1%)                 0                     0       
#>   Reoviral infection                                                             0                     0                  1 (5.6%)   
#>   Schistosomiasis bladder                                                     1 (9.1%)                 0                     0       
#>   Splenic infection helminthic                                                   0                  1 (4.8%)                 0       
#>   Systemic mycosis                                                               0                     0                  1 (5.6%)   
#>   Thyroid echinococciasis                                                        0                  1 (4.8%)                 0       
#>   Ureaplasma infection                                                           0                  1 (4.8%)                 0       
#>   Urinary tract infection bacterial                                              0                  1 (4.8%)                 0       
#>   Vulvovaginitis helminthic                                                      0                  1 (4.8%)                 0       
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of patients with at least one adverse event                   3 (27.3%)             8 (38.1%)             6 (33.3%)   
#>   Total number of events                                                         4                     8                     7       
#>   Acne occupational                                                           1 (9.1%)                 0                     0       
#>   Acral peeling skin syndrome                                                 1 (9.1%)                 0                     0       
#>   Eczema asteatotic                                                              0                  1 (4.8%)                 0       
#>   Erythema ab igne                                                               0                     0                  1 (5.6%)   
#>   Hair growth rate abnormal                                                   1 (9.1%)                 0                     0       
#>   Hyper IgD syndrome                                                             0                     0                  1 (5.6%)   
#>   Hypokeratosis                                                                  0                  1 (4.8%)                 0       
#>   Injection site telangiectasia                                                  0                     0                  1 (5.6%)   
#>   Lip oedema                                                                     0                     0                  1 (5.6%)   
#>   Measles                                                                     1 (9.1%)                 0                     0       
#>   Nail pterygium                                                                 0                     0                  1 (5.6%)   
#>   Netherton's syndrome                                                           0                  1 (4.8%)                 0       
#>   Palpable purpura                                                               0                  1 (4.8%)                 0       
#>   Pruritus                                                                       0                  1 (4.8%)                 0       
#>   Rash rubelliform                                                               0                  1 (4.8%)                 0       
#>   Sea bather's eruption                                                          0                  1 (4.8%)                 0       
#>   Solar lentigo                                                                  0                     0                  1 (5.6%)   
#>   Vulvovaginal pruritus                                                          0                     0                  1 (5.6%)   
#>   Waardenburg syndrome                                                           0                  1 (4.8%)                 0       
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of patients with at least one adverse event                   2 (18.2%)             8 (38.1%)             6 (33.3%)   
#>   Total number of events                                                         3                     10                    8       
#>   Cardiac lymphangioma                                                           0                     0                  1 (5.6%)   
#>   Cervix adenomatous polyp                                                       0                     0                  1 (5.6%)   
#>   Colorectal cancer stage IV                                                     0                  1 (4.8%)                 0       
#>   Endocrine neoplasm malignant                                                   0                     0                  1 (5.6%)   
#>   Fibrosarcoma metastatic                                                        0                  1 (4.8%)                 0       
#>   Gallbladder adenoma                                                            0                  1 (4.8%)                 0       
#>   Gallbladder papilloma                                                          0                  1 (4.8%)                 0       
#>   Glomus jugulare tumour                                                         0                     0                  1 (5.6%)   
#>   Hereditary non-polyposis colorectal cancer syndrome                            0                  1 (4.8%)                 0       
#>   Hodgkin's disease mixed cellularity refractory                              1 (9.1%)                 0                     0       
#>   Lung carcinoma cell type unspecified stage 0                                   0                  1 (4.8%)                 0       
#>   Malignant bowel obstruction                                                 1 (9.1%)                 0                     0       
#>   Malignant melanoma                                                          1 (9.1%)                 0                     0       
#>   Metastases to uterus                                                           0                  1 (4.8%)                 0       
#>   Mismatch repair cancer syndrome                                                0                     0                  1 (5.6%)   
#>   Ovarian germ cell polyembryoma stage II                                        0                  1 (4.8%)                 0       
#>   Serous cystadenocarcinoma ovary                                                0                     0                  1 (5.6%)   
#>   Small intestine carcinoma stage II                                             0                  1 (4.8%)                 0       
#>   T-cell lymphoma recurrent                                                      0                     0                  1 (5.6%)   
#>   Thyroid neoplasm                                                               0                     0                  1 (5.6%)   
#>   Tumour pain                                                                    0                  1 (4.8%)                 0       
#> 
#> Surgical and medical procedures
#>   Total number of patients with at least one adverse event                   4 (36.4%)             8 (38.1%)             4 (22.2%)   
#>   Total number of events                                                         4                     10                    4       
#>   Salpingo-oophorectomy bilateral                                                0                     0                 2 (11.1%)   
#>   Arterial catheterisation                                                       0                  1 (4.8%)                 0       
#>   Breast reconstruction                                                          0                  1 (4.8%)                 0       
#>   Duodeno-jejunal bypass sleeve therapy                                          0                  1 (4.8%)                 0       
#>   External ear lesion excision                                                   0                  1 (4.8%)                 0       
#>   Local anaesthesia                                                              0                  1 (4.8%)                 0       
#>   Meningeal repair                                                               0                  1 (4.8%)                 0       
#>   Muscle relaxant therapy                                                        0                  1 (4.8%)                 0       
#>   Oesophageal lesion excision                                                 1 (9.1%)                 0                     0       
#>   Percutaneous ethanol injection therapy                                         0                  1 (4.8%)                 0       
#>   Pericardiotomy                                                              1 (9.1%)                 0                     0       
#>   Phlebotomy                                                                     0                  1 (4.8%)                 0       
#>   Premedication                                                               1 (9.1%)                 0                     0       
#>   Resuscitation                                                                  0                     0                  1 (5.6%)   
#>   Transgender operation                                                       1 (9.1%)                 0                     0       
#>   Tubal insufflation                                                             0                     0                  1 (5.6%)   
#>   Wound drainage                                                                 0                  1 (4.8%)                 0       
#> 
#> Gastrointestinal disorders
#>   Total number of patients with at least one adverse event                   3 (27.3%)             7 (33.3%)             3 (16.7%)   
#>   Total number of events                                                         4                     7                     3       
#>   Amoebic colitis                                                             1 (9.1%)                 0                     0       
#>   Anal neoplasm                                                               1 (9.1%)                 0                     0       
#>   Campylobacter colitis                                                          0                  1 (4.8%)                 0       
#>   Currarino syndrome                                                             0                  1 (4.8%)                 0       
#>   Cytomegalovirus enteritis                                                      0                  1 (4.8%)                 0       
#>   Gastric cancer stage 0                                                         0                  1 (4.8%)                 0       
#>   Hypercementosis                                                                0                  1 (4.8%)                 0       
#>   Hypogeusia                                                                     0                  1 (4.8%)                 0       
#>   Oesophageal carcinoma stage 0                                                  0                  1 (4.8%)                 0       
#>   Premenstrual cramps                                                            0                     0                  1 (5.6%)   
#>   Pyogenic granuloma                                                          1 (9.1%)                 0                     0       
#>   Spigelian hernia                                                               0                     0                  1 (5.6%)   
#>   Syphilis anal                                                                  0                     0                  1 (5.6%)   
#>   Vitello-intestinal duct remnant                                             1 (9.1%)                 0                     0       
#> 
#> Blood and lymphatic system disorders
#>   Total number of patients with at least one adverse event                   4 (36.4%)              4 (19%)              4 (22.2%)   
#>   Total number of events                                                         7                     4                     4       
#>   Anaemic hypoxia                                                             1 (9.1%)                 0                     0       
#>   Cold type haemolytic anaemia                                                   0                     0                  1 (5.6%)   
#>   Congenital dysfibrinogenaemia                                                  0                     0                  1 (5.6%)   
#>   Erythrosis                                                                     0                  1 (4.8%)                 0       
#>   Factor IX inhibition                                                        1 (9.1%)                 0                     0       
#>   Follicular dendritic cell sarcoma                                           1 (9.1%)                 0                     0       
#>   High grade B-cell lymphoma Burkitt-like lymphoma recurrent                     0                  1 (4.8%)                 0       
#>   Lymphadenitis viral                                                            0                  1 (4.8%)                 0       
#>   Lymphoid tissue hypoplasia                                                     0                     0                  1 (5.6%)   
#>   Non-Hodgkin's lymphoma stage III                                               0                  1 (4.8%)                 0       
#>   Non-immune heparin associated thrombocytopenia                              1 (9.1%)                 0                     0       
#>   Post transfusion purpura                                                    1 (9.1%)                 0                     0       
#>   Precursor B-lymphoblastic lymphoma                                          1 (9.1%)                 0                     0       
#>   Splenic infarction                                                             0                     0                  1 (5.6%)   
#>   Splenic marginal zone lymphoma                                              1 (9.1%)                 0                     0       
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of patients with at least one adverse event                    1 (9.1%)             3 (14.3%)             6 (33.3%)   
#>   Total number of events                                                         1                     3                     6       
#>   Camptodactyly congenital                                                       0                     0                  1 (5.6%)   
#>   Compression fracture                                                           0                     0                  1 (5.6%)   
#>   Congenital kyphoscoliosis                                                   1 (9.1%)                 0                     0       
#>   Infrapatellar fat pad inflammation                                             0                     0                  1 (5.6%)   
#>   Kyphoscoliosis                                                                 0                     0                  1 (5.6%)   
#>   Kyphosis radiation                                                             0                  1 (4.8%)                 0       
#>   Lordosis                                                                       0                  1 (4.8%)                 0       
#>   Musculoskeletal chest pain                                                     0                     0                  1 (5.6%)   
#>   Staphylococcal osteomyelitis                                                   0                  1 (4.8%)                 0       
#>   Temporomandibular joint syndrome                                               0                     0                  1 (5.6%)   
#> 
#> Vascular disorders
#>   Total number of patients with at least one adverse event                   2 (18.2%)              4 (19%)              4 (22.2%)   
#>   Total number of events                                                         2                     4                     6       
#>   Adrenal thrombosis                                                             0                     0                  1 (5.6%)   
#>   Femoral vein perforation                                                       0                     0                  1 (5.6%)   
#>   Hepatic vein embolism                                                          0                  1 (4.8%)                 0       
#>   Hoigne's syndrome                                                              0                  1 (4.8%)                 0       
#>   Iliac artery embolism                                                          0                     0                  1 (5.6%)   
#>   Intracranial arterial fenestration                                          1 (9.1%)                 0                     0       
#>   Parkes-Weber syndrome                                                          0                     0                  1 (5.6%)   
#>   Pharyngeal haematoma                                                           0                  1 (4.8%)                 0       
#>   Portal vein thrombosis                                                         0                     0                  1 (5.6%)   
#>   Post cardiac arrest syndrome                                                1 (9.1%)                 0                     0       
#>   Stroke in evolution                                                            0                  1 (4.8%)                 0       
#>   Superior vena cava syndrome                                                    0                     0                  1 (5.6%)   
#> 
#> Nervous system disorders
#>   Total number of patients with at least one adverse event                   4 (36.4%)             3 (14.3%)             2 (11.1%)   
#>   Total number of events                                                         5                     3                     3       
#>   Baltic myoclonic epilepsy                                                      0                  1 (4.8%)                 0       
#>   Brow ptosis                                                                    0                  1 (4.8%)                 0       
#>   Bulbar poliomyelitis                                                        1 (9.1%)                 0                     0       
#>   Cerebrospinal fistula                                                       1 (9.1%)                 0                     0       
#>   Congenital hemiparesis                                                      1 (9.1%)                 0                     0       
#>   Cornelia de Lange syndrome                                                     0                     0                  1 (5.6%)   
#>   Creutzfeldt-Jakob disease                                                      0                     0                  1 (5.6%)   
#>   Moebius II syndrome                                                            0                  1 (4.8%)                 0       
#>   Postictal state                                                             1 (9.1%)                 0                     0       
#>   Psychomotor hyperactivity                                                      0                     0                  1 (5.6%)   
#>   West Nile viral infection                                                   1 (9.1%)                 0                     0       
#> 
#> Reproductive system and breast disorders
#>   Total number of patients with at least one adverse event                   4 (36.4%)             3 (14.3%)             2 (11.1%)   
#>   Total number of events                                                         4                     3                     2       
#>   Genital injury                                                              1 (9.1%)                 0                     0       
#>   Haemangioma of breast                                                       1 (9.1%)                 0                     0       
#>   Mammary fistula                                                                0                     0                  1 (5.6%)   
#>   Neuroendocrine breast tumour                                                1 (9.1%)                 0                     0       
#>   Organic erectile dysfunction                                                   0                  1 (4.8%)                 0       
#>   Ovarian germ cell teratoma stage I                                             0                  1 (4.8%)                 0       
#>   Phyllodes tumour                                                               0                  1 (4.8%)                 0       
#>   Uterine cervical pain                                                       1 (9.1%)                 0                     0       
#>   Uterine pain                                                                   0                     0                  1 (5.6%)   
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of patients with at least one adverse event                   2 (18.2%)             5 (23.8%)             2 (11.1%)   
#>   Total number of events                                                         3                     5                     3       
#>   Atelectasis                                                                    0                     0                  1 (5.6%)   
#>   Dyspnoea exertional                                                            0                     0                  1 (5.6%)   
#>   Loeffler's syndrome                                                            0                  1 (4.8%)                 0       
#>   Nasal mucosal ulcer                                                            0                  1 (4.8%)                 0       
#>   Oesophagobronchial fistula                                                  1 (9.1%)                 0                     0       
#>   PFAPA syndrome                                                                 0                  1 (4.8%)                 0       
#>   Pulmonary contusion                                                            0                     0                  1 (5.6%)   
#>   Pulmonary sequestration                                                     1 (9.1%)                 0                     0       
#>   Sinus polyp                                                                    0                  1 (4.8%)                 0       
#>   Squamous cell carcinoma of the hypopharynx                                     0                  1 (4.8%)                 0       
#>   Tracheal stenosis                                                           1 (9.1%)                 0                     0       
#> 
#> Congenital, familial and genetic disorders
#>   Total number of patients with at least one adverse event                   2 (18.2%)              4 (19%)              2 (11.1%)   
#>   Total number of events                                                         2                     6                     2       
#>   Ataxia telangiectasia                                                       1 (9.1%)                 0                     0       
#>   Bladder agenesis                                                            1 (9.1%)                 0                     0       
#>   Cerebral atrophy congenital                                                    0                     0                  1 (5.6%)   
#>   Cobb syndrome                                                                  0                     0                  1 (5.6%)   
#>   Congenital cutis laxa                                                          0                  1 (4.8%)                 0       
#>   Familial amyloidosis                                                           0                  1 (4.8%)                 0       
#>   Gaucher's disease type II                                                      0                  1 (4.8%)                 0       
#>   Haemoglobin E disease                                                          0                  1 (4.8%)                 0       
#>   Mitochondrial hepatopathy                                                      0                  1 (4.8%)                 0       
#>   Wolfram syndrome                                                               0                  1 (4.8%)                 0       
#> 
#> Eye disorders
#>   Total number of patients with at least one adverse event                       0                  4 (19%)              4 (22.2%)   
#>   Total number of events                                                         0                     4                     4       
#>   Chromatopsia                                                                   0                     0                  1 (5.6%)   
#>   Eversion of lacrimal punctum                                                   0                     0                  1 (5.6%)   
#>   Haemangioma of retina                                                          0                  1 (4.8%)                 0       
#>   Macular rupture                                                                0                     0                  1 (5.6%)   
#>   Marfan's syndrome                                                              0                  1 (4.8%)                 0       
#>   Neurologic neglect syndrome                                                    0                     0                  1 (5.6%)   
#>   Pupillary deformity                                                            0                  1 (4.8%)                 0       
#>   Retinal melanoma                                                               0                  1 (4.8%)                 0       
#> 
#> General disorders and administration site conditions
#>   Total number of patients with at least one adverse event                       0                  4 (19%)              3 (16.7%)   
#>   Total number of events                                                         0                     5                     4       
#>   Delirium febrile                                                               0                  1 (4.8%)                 0       
#>   Enzyme inhibition                                                              0                  1 (4.8%)                 0       
#>   General physical health deterioration                                          0                  1 (4.8%)                 0       
#>   Ill-defined disorder                                                           0                     0                  1 (5.6%)   
#>   Multiple organ dysfunction syndrome                                            0                  1 (4.8%)                 0       
#>   Puncture site induration                                                       0                     0                  1 (5.6%)   
#>   Soft tissue inflammation                                                       0                  1 (4.8%)                 0       
#>   Temperature regulation disorder                                                0                     0                  1 (5.6%)   
#>   Vaccination site reaction                                                      0                     0                  1 (5.6%)   
#> 
#> Renal and urinary disorders
#>   Total number of patients with at least one adverse event                       0                  4 (19%)              2 (11.1%)   
#>   Total number of events                                                         0                     4                     2       
#>   Bladder diverticulum                                                           0                  1 (4.8%)                 0       
#>   Cystitis-like symptom                                                          0                  1 (4.8%)                 0       
#>   Persistent urogenital sinus                                                    0                     0                  1 (5.6%)   
#>   Transitional cell cancer of the renal pelvis and ureter regional               0                  1 (4.8%)                 0       
#>   Tubulointerstitial nephritis and uveitis syndrome                              0                     0                  1 (5.6%)   
#>   Urogenital prolapse                                                            0                  1 (4.8%)                 0       
#> 
#> Pregnancy, puerperium and perinatal conditions
#>   Total number of patients with at least one adverse event                    1 (9.1%)             3 (14.3%)              1 (5.6%)   
#>   Total number of events                                                         1                     3                     2       
#>   Congenital hepatitis B infection                                               0                  1 (4.8%)                 0       
#>   Cranial sutures widening                                                    1 (9.1%)                 0                     0       
#>   Foetal malformation                                                            0                  1 (4.8%)                 0       
#>   Oligohydramnios                                                                0                     0                  1 (5.6%)   
#>   Placenta accreta                                                               0                  1 (4.8%)                 0       
#>   Risk of future pregnancy miscarriage                                           0                     0                  1 (5.6%)   
#> 
#> Cardiac disorders
#>   Total number of patients with at least one adverse event                    1 (9.1%)             3 (14.3%)                 0       
#>   Total number of events                                                         1                     3                     0       
#>   Foetal arrhythmia                                                           1 (9.1%)                 0                     0       
#>   Oedema due to cardiac disease                                                  0                  1 (4.8%)                 0       
#>   Pericardial effusion                                                           0                  1 (4.8%)                 0       
#>   Right ventricular enlargement                                                  0                  1 (4.8%)                 0       
#> 
#> Metabolism and nutrition disorders
#>   Total number of patients with at least one adverse event                    1 (9.1%)              2 (9.5%)                 0       
#>   Total number of events                                                         1                     2                     0       
#>   Infantile scurvy                                                               0                  1 (4.8%)                 0       
#>   Osteoporotic fracture                                                          0                  1 (4.8%)                 0       
#>   Pseudohypoparathyroidism                                                    1 (9.1%)                 0                     0       
#> 
#> Product issues
#>   Total number of patients with at least one adverse event                       0                  1 (4.8%)             2 (11.1%)   
#>   Total number of events                                                         0                     1                     2       
#>   Product closure issue                                                          0                     0                  1 (5.6%)   
#>   Product outer packaging issue                                                  0                  1 (4.8%)                 0       
#>   Product packaging counterfeit                                                  0                     0                  1 (5.6%)   
#> 
#> Immune system disorders
#>   Total number of patients with at least one adverse event                       0                  1 (4.8%)              1 (5.6%)   
#>   Total number of events                                                         0                     1                     1       
#>   Corneal graft rejection                                                        0                     0                  1 (5.6%)   
#>   Cutaneous malacoplakia                                                         0                  1 (4.8%)                 0       
#> 
#> Social circumstances
#>   Total number of patients with at least one adverse event                       0                  1 (4.8%)              1 (5.6%)   
#>   Total number of events                                                         0                     1                     1       
#>   Early adult transition                                                         0                     0                  1 (5.6%)   
#>   Suicide of relative                                                            0                  1 (4.8%)                 0       
#> 
#> Ear and labyrinth disorders
#>   Total number of patients with at least one adverse event                    1 (9.1%)                 0                     0       
#>   Total number of events                                                         1                     0                     0       
#>   Deaf mutism                                                                 1 (9.1%)                 0                     0       
#> 
#> Endocrine disorders
#>   Total number of patients with at least one adverse event                       0                     0                  1 (5.6%)   
#>   Total number of events                                                         0                     0                     1       
#>   Thyroid cancer                                                                 0                     0                  1 (5.6%)   
#> 
#> Hepatobiliary disorders
#>   Total number of patients with at least one adverse event                       0                  1 (4.8%)                 0       
#>   Total number of events                                                         0                     1                     0       
#>   Hepatic echinococciasis                                                        0                  1 (4.8%)                 0       
#> 
#> Psychiatric disorders
#>   Total number of patients with at least one adverse event                       0                     0                  1 (5.6%)   
#>   Total number of events                                                         0                     0                     1       
#>   Abulia                                                                         0                     0                  1 (5.6%)