Create an AE Overview Summary Table (AET01)

t_ae_oview returns adverse events according to STREAM format AET01

t_ae_oview(id, class, term, dthfl, dcsreas, aesdth, aeser, aeacn, arel, aerel,
  aetoxgr, col_by, total = "All Patients")

Arguments

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

class
system organ class variable.

term
preferred term variable.

dthfl
subject death flag variable.

dcsreas
reason for discontinuation from study variable.

aesdth
results in death variable.

aeser
serious event variable.

aeacn
action taken with study treatment variable.

arel
analysis causality variable.

aerel
causality variable.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing.

total
character string that will be used as a label for a column with pooled total population, default here is "All Patients".

atoxgr
analysis toxicity grade variable.

Value

rtable object

Examples

library(dplyr)
suppressPackageStartupMessages(library(tidyverse))
library(rtables)

#> Warning: package ‘rtables’ was built under R version 3.4.4

library(haven)
adae = read_sas("/opt/bee/home_nas/zhanc107/osprey/adam/adae.sas7bdat")

#> Error: '/opt/bee/home_nas/zhanc107/osprey/adam/adae.sas7bdat' does not exist.

tbl <- t_ae_oview(
   id = adae$USUBJID,
   class = adae$AESOC,
   term = adae$AEDECOD,
   dthfl = adae$DTHFL,
   dcsreas = adae$DCSREAS,
   aesdth = adae$AESDTH,
   aeser = adae$AESER,
   aeacn = adae$AEACN,
   arel = adae$AREL,
   aerel = adae$AEREL,
   aetoxgr = adae$AETOXGR,
   col_by = factor(adae$ARM),
   total="All Patients"
)

#> Error in check_col_by(col_by, min_num_levels = 1): could not find function "check_col_by"

tbl

#> function (src, ...) 
#> {
#>     UseMethod("tbl")
#> }
#> <environment: namespace:dplyr>