Create a Patient Disposition (DST01) Table

t_ds returns patient disposition according to STREAM format DST01

t_ds(class, term, sub = NULL, id, col_by, total = "All Patients", ...)

Arguments

term
reason for discontinuation from study.

sub
dataframe of additional subsets of terms

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing.

total
character string that will be used as a label for a column with pooled total population, default here is "All Patients", if set to "NONE" then the "All Patients" column is suppressed.

Value

rtable object

Examples

# Simple example
library(tibble)
library(dplyr)
library(rtables)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)

ae_lookup <- tribble(
~CLASS,         ~TERM,   ~GRADE,
"cl A",   "trm A_1/2",        1,
"cl A",   "trm A_2/2",        2,
"cl B",   "trm B_1/3",        2,
"cl B",   "trm B_2/3",        3,
"cl B",   "trm B_3/3",        1,
"cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ds(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients",
)

tbl

#>                       ARM A               ARM B            All Patients
#>                       (N=3)               (N=7)               (N=10)   
#> -----------------------------------------------------------------------
#> cl A                1 (33.33%)          3 (42.86%)           4 (40%)   
#> 
#>   trm A_1/2         1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#>   trm A_2/2         1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#> cl B                1 (33.33%)          3 (42.86%)           4 (40%)   
#> 
#>   trm B_1/3             0               1 (14.29%)           1 (10%)   
#> 
#>   trm B_2/3         1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#>   trm B_3/3             0               1 (14.29%)           1 (10%)   
#> 
#> cl C                1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#>   trm C_1/1         1 (33.33%)          1 (14.29%)           2 (20%)   

tbl2 <- t_ds(
  class = ANL$CLASS,
  term =  ANL$TERM,
  sub = data.frame(GRADE = ANL$GRADE),
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients",
)

tbl2

#>                       ARM A               ARM B            All Patients
#>                       (N=3)               (N=7)               (N=10)   
#> -----------------------------------------------------------------------
#> cl A                1 (33.33%)          3 (42.86%)           4 (40%)   
#> 
#>   trm A_1/2         1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#>     1               1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#>   trm A_2/2         1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#>     2               1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#> cl B                1 (33.33%)          3 (42.86%)           4 (40%)   
#> 
#>   trm B_1/3             0               1 (14.29%)           1 (10%)   
#> 
#>     2                   0               1 (14.29%)           1 (10%)   
#> 
#>   trm B_2/3         1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#>     3               1 (33.33%)          2 (28.57%)           3 (30%)   
#> 
#>   trm B_3/3             0               1 (14.29%)           1 (10%)   
#> 
#>     1                   0               1 (14.29%)           1 (10%)   
#> 
#> cl C                1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#>   trm C_1/1         1 (33.33%)          1 (14.29%)           2 (20%)   
#> 
#>     1               1 (33.33%)          1 (14.29%)           2 (20%)   

# test wth Ninas study data
#asl <- read.bce("/opt/BIOSTAT/prod/s30103j/libraries/asl.sas7bdat")
asl <- read.bce("/opt/BIOSTAT/home/qit3/cdt70094/s30103j/libraries/aae_b.sas7bdat")

#> Error in read.bce("/opt/BIOSTAT/home/qit3/cdt70094/s30103j/libraries/aae_b.sas7bdat"): could not find function "read.bce"

asl <- asl %>% filter(AEREL1 != "" & AESOC != "" & AEDECOD != "")# & (AESOC == "INVESTIGATIONS" | AESOC == "CARDIAC DISORDERS"))

#> Error in eval(lhs, parent, parent): object 'asl' not found

asl <- asl[1:20,]

#> Error in eval(expr, envir, enclos): object 'asl' not found

tbl3 <- t_ds(
  class = asl$AESOC,
  term =  asl$AEREL1,
  sub = data.frame(AEDECOD = asl$AEDECOD),
  id = asl$USUBJID,
  col_by = factor(asl$ARMCD),
  total = "All Patients",
)

#> Error in factor(asl$ARMCD): object 'asl' not found

tbl3

#> Error in eval(expr, envir, enclos): object 'tbl3' not found