Create an Adverse Events (AET02) Table

t_ae returns adverse events according to STREAM format AET02

t_ae(class, term, id, col_by, total = "All Patients", sort_by = "count",
  ...)

Arguments

class
system organ class variable.

term
preferred term variable.

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing.

total
character string that will be used as a label for a column with pooled total population, default here is "All Patients", if set to "NONE" then the "All Patients" column is suppressed.

sort_by
character string that defines the ordering of the class and term variables in the output table, options: "alphabetical" or "count", default here is set to "count"

Value

rtable object

Examples

# Simple example
library(tibble)
library(dplyr)
library(rtables)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)

ae_lookup <- tribble(
~CLASS,         ~TERM,   ~GRADE,
"cl A",   "trm A_1/2",        1,
"cl A",   "trm A_2/2",        2,
"cl B",   "trm B_1/3",        2,
"cl B",   "trm B_2/3",        3,
"cl B",   "trm B_3/3",        1,
"cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ae(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients",
)

tbl

#> MedDRA System Organ Class                                            ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                              (N=3)               (N=7)               (N=10)   
#> ----------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event           2 (66.7%)           4 (57.1%)            6 (60%)   
#> 
#> Overall total number of events                                         5                   9                   14     
#> cl A
#>   Total number of patients with at least one adverse event         1 (33.3%)           3 (42.9%)            4 (40%)   
#> 
#>   Total number of events                                               3                   4                   7      
#> 
#>   trm A_2/2                                                        1 (33.3%)           2 (28.6%)            3 (30%)   
#> 
#>   trm A_1/2                                                        1 (33.3%)           1 (14.3%)            2 (20%)   
#> 
#> cl B
#>   Total number of patients with at least one adverse event         1 (33.3%)           3 (42.9%)            4 (40%)   
#> 
#>   Total number of events                                               1                   4                   5      
#> 
#>   trm B_2/3                                                        1 (33.3%)           2 (28.6%)            3 (30%)   
#> 
#>   trm B_1/3                                                            0               1 (14.3%)            1 (10%)   
#> 
#>   trm B_3/3                                                            0               1 (14.3%)            1 (10%)   
#> 
#> cl C
#>   Total number of patients with at least one adverse event         1 (33.3%)           1 (14.3%)            2 (20%)   
#> 
#>   Total number of events                                               1                   1                   2      
#> 
#>   trm C_1/1                                                        1 (33.3%)           1 (14.3%)            2 (20%)   

# Simple example 2

library(dplyr)
library(rtables)
library(random.cdisc.data)

data <- left_join(radam("AAE", N=10),radam("ADSL", N=10))

#> Joining, by = c("STUDYID", "USUBJID")

tbl4 <- t_ae(
  class = data$AEBODSYS,
  term =  data$AEDECOD,
  id = data$USUBJID,
  col_by = factor(data$ARM),
  total = "None",
  sort_by = "alphabetical"
)

tbl4

#> MedDRA System Organ Class                                                    ARM A            ARM B            None   
#>   MedDRA Preferred Term                                                      (N=5)            (N=5)           (N=10)  
#> ----------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event                   5 (100%)         5 (100%)         10 (100%)
#> 
#> Overall total number of events                                                34               20               54    
#> Blood and lymphatic system disorders
#>   Total number of patients with at least one adverse event                     0             1 (20%)          1 (10%) 
#> 
#>   Total number of events                                                       0                1                1    
#> 
#>   Blood disorder                                                               0             1 (20%)          1 (10%) 
#> 
#> Cardiac disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Tricuspid valve sclerosis                                                 1 (20%)             0             1 (10%) 
#> 
#> Congenital, familial and genetic disorders
#>   Total number of patients with at least one adverse event                  3 (60%)          1 (20%)          4 (40%) 
#> 
#>   Total number of events                                                       4                1                5    
#> 
#>   Congenital aortic dilatation                                                 0             1 (20%)          1 (10%) 
#> 
#>   Congenital cystic disease of liver                                        1 (20%)             0             1 (10%) 
#> 
#>   Craniotabes                                                               1 (20%)             0             1 (10%) 
#> 
#>   Enteric duplication                                                       1 (20%)             0             1 (10%) 
#> 
#>   Tibial agenesis                                                           1 (20%)             0             1 (10%) 
#> 
#> Endocrine disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Emphysematous pyelonephritis                                              1 (20%)             0             1 (10%) 
#> 
#> Gastrointestinal disorders
#>   Total number of patients with at least one adverse event                     0             1 (20%)          1 (10%) 
#> 
#>   Total number of events                                                       0                1                1    
#> 
#>   Abscess intestinal                                                           0             1 (20%)          1 (10%) 
#> 
#> General disorders and administration site conditions
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Catheter site oedema                                                         0             1 (20%)          1 (10%) 
#> 
#>   Membrane stabilising effect                                               1 (20%)             0             1 (10%) 
#> 
#> Immune system disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Dermatomyositis                                                              0             1 (20%)          1 (10%) 
#> 
#>   Myasthenia gravis crisis                                                  1 (20%)             0             1 (10%) 
#> 
#> Infections and infestations
#>   Total number of patients with at least one adverse event                  3 (60%)          1 (20%)          4 (40%) 
#> 
#>   Total number of events                                                       5                1                6    
#> 
#>   Gallbladder empyema                                                       1 (20%)             0             1 (10%) 
#> 
#>   Mucormycosis                                                              1 (20%)             0             1 (10%) 
#> 
#>   Pleural infection                                                         1 (20%)             0             1 (10%) 
#> 
#>   Retroperitoneal infection                                                 1 (20%)             0             1 (10%) 
#> 
#>   Tuberculosis of genitourinary system                                         0             1 (20%)          1 (10%) 
#> 
#>   Viral rash                                                                1 (20%)             0             1 (10%) 
#> 
#> Injury, poisoning and procedural complications
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Instillation site oedema                                                  1 (20%)             0             1 (10%) 
#> 
#>   Ligament rupture                                                             0             1 (20%)          1 (10%) 
#> 
#> Investigations
#>   Total number of patients with at least one adverse event                  3 (60%)          2 (40%)          5 (50%) 
#> 
#>   Total number of events                                                       5                4                9    
#> 
#>   Base excess positive                                                      1 (20%)             0             1 (10%) 
#> 
#>   Biopsy small intestine abnormal                                              0             1 (20%)          1 (10%) 
#> 
#>   Biopsy tongue normal                                                      1 (20%)             0             1 (10%) 
#> 
#>   Crossmatch incompatible                                                      0             1 (20%)          1 (10%) 
#> 
#>   Enzyme level test                                                         1 (20%)             0             1 (10%) 
#> 
#>   Human herpes virus 8 test                                                    0             1 (20%)          1 (10%) 
#> 
#>   Skin test negative                                                        1 (20%)             0             1 (10%) 
#> 
#>   Urine copper                                                                 0             1 (20%)          1 (10%) 
#> 
#>   Urobilinogen faeces normal                                                1 (20%)             0             1 (10%) 
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Muscle contusion                                                          1 (20%)             0             1 (10%) 
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of patients with at least one adverse event                  3 (60%)          2 (40%)          5 (50%) 
#> 
#>   Total number of events                                                       4                2                6    
#> 
#>   Adult T-cell lymphoma/leukaemia stage IV                                     0             1 (20%)          1 (10%) 
#> 
#>   Congenital melanocytic naevus                                                0             1 (20%)          1 (10%) 
#> 
#>   Liver carcinoma ruptured                                                  1 (20%)             0             1 (10%) 
#> 
#>   Malignant neoplasm of thymus                                              1 (20%)             0             1 (10%) 
#> 
#>   Paranasal sinus and nasal cavity malignant neoplasm stage 0               1 (20%)             0             1 (10%) 
#> 
#>   Urinary bladder polyp                                                     1 (20%)             0             1 (10%) 
#> 
#> Nervous system disorders
#>   Total number of patients with at least one adverse event                  4 (80%)          1 (20%)          5 (50%) 
#> 
#>   Total number of events                                                       4                1                5    
#> 
#>   Cerebral reperfusion injury                                               1 (20%)             0             1 (10%) 
#> 
#>   Choriomeningitis lymphocytic                                              1 (20%)             0             1 (10%) 
#> 
#>   Excessive eye blinking                                                       0             1 (20%)          1 (10%) 
#> 
#>   Exertional headache                                                       1 (20%)             0             1 (10%) 
#> 
#>   Prosopagnosia                                                             1 (20%)             0             1 (10%) 
#> 
#> Psychiatric disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Somatic symptom disorder                                                  1 (20%)             0             1 (10%) 
#> 
#>   Tourette's disorder                                                          0             1 (20%)          1 (10%) 
#> 
#> Renal and urinary disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Alkaptonuria                                                              1 (20%)             0             1 (10%) 
#> 
#> Reproductive system and breast disorders
#>   Total number of patients with at least one adverse event                     0             1 (20%)          1 (10%) 
#> 
#>   Total number of events                                                       0                1                1    
#> 
#>   Bleeding anovulatory                                                         0             1 (20%)          1 (10%) 
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of patients with at least one adverse event                     0             1 (20%)          1 (10%) 
#> 
#>   Total number of events                                                       0                1                1    
#> 
#>   Hypopharyngeal cancer stage III                                              0             1 (20%)          1 (10%) 
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of patients with at least one adverse event                  2 (40%)          2 (40%)          4 (40%) 
#> 
#>   Total number of events                                                       2                2                4    
#> 
#>   Harlequin syndrome                                                           0             1 (20%)          1 (10%) 
#> 
#>   Nail pitting                                                              1 (20%)             0             1 (10%) 
#> 
#>   Skin sensitisation                                                           0             1 (20%)          1 (10%) 
#> 
#>   Vaccination site eczema                                                   1 (20%)             0             1 (10%) 
#> 
#> Surgical and medical procedures
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Colporrhaphy                                                              1 (20%)             0             1 (10%) 
#> 
#> Vascular disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Instillation site thrombosis                                                 0             1 (20%)          1 (10%) 
#> 
#>   Necrotising herpetic retinopathy                                          1 (20%)             0             1 (10%) 

data <- read_bce("/opt/BIOSTAT/home_ext2/qit3/cdt70194/go39733/libraries/adae.sas7bdat")

#> Error in read_bce("/opt/BIOSTAT/home_ext2/qit3/cdt70194/go39733/libraries/adae.sas7bdat"): could not find function "read_bce"

tbl5 <- t_ae(
  class = data$AEBODSYS,
  term =  data$AEDECOD,
  id = data$USUBJID,
  col_by = factor(data$ARM)
)

tbl5

#> MedDRA System Organ Class                                                     ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                                       (N=5)               (N=5)               (N=10)   
#> -------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event                     5 (100%)            5 (100%)           10 (100%)  
#> 
#> Overall total number of events                                                  34                  20                  54     
#> Investigations
#>   Total number of patients with at least one adverse event                   3 (60%)             2 (40%)             5 (50%)   
#> 
#>   Total number of events                                                        5                   4                   9      
#> 
#>   Base excess positive                                                       1 (20%)                0                1 (10%)   
#> 
#>   Biopsy small intestine abnormal                                               0                1 (20%)             1 (10%)   
#> 
#>   Biopsy tongue normal                                                       1 (20%)                0                1 (10%)   
#> 
#>   Crossmatch incompatible                                                       0                1 (20%)             1 (10%)   
#> 
#>   Enzyme level test                                                          1 (20%)                0                1 (10%)   
#> 
#>   Human herpes virus 8 test                                                     0                1 (20%)             1 (10%)   
#> 
#>   Skin test negative                                                         1 (20%)                0                1 (10%)   
#> 
#>   Urine copper                                                                  0                1 (20%)             1 (10%)   
#> 
#>   Urobilinogen faeces normal                                                 1 (20%)                0                1 (10%)   
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of patients with at least one adverse event                   3 (60%)             2 (40%)             5 (50%)   
#> 
#>   Total number of events                                                        4                   2                   6      
#> 
#>   Adult T-cell lymphoma/leukaemia stage IV                                      0                1 (20%)             1 (10%)   
#> 
#>   Congenital melanocytic naevus                                                 0                1 (20%)             1 (10%)   
#> 
#>   Liver carcinoma ruptured                                                   1 (20%)                0                1 (10%)   
#> 
#>   Malignant neoplasm of thymus                                               1 (20%)                0                1 (10%)   
#> 
#>   Paranasal sinus and nasal cavity malignant neoplasm stage 0                1 (20%)                0                1 (10%)   
#> 
#>   Urinary bladder polyp                                                      1 (20%)                0                1 (10%)   
#> 
#> Nervous system disorders
#>   Total number of patients with at least one adverse event                   4 (80%)             1 (20%)             5 (50%)   
#> 
#>   Total number of events                                                        4                   1                   5      
#> 
#>   Cerebral reperfusion injury                                                1 (20%)                0                1 (10%)   
#> 
#>   Choriomeningitis lymphocytic                                               1 (20%)                0                1 (10%)   
#> 
#>   Excessive eye blinking                                                        0                1 (20%)             1 (10%)   
#> 
#>   Exertional headache                                                        1 (20%)                0                1 (10%)   
#> 
#>   Prosopagnosia                                                              1 (20%)                0                1 (10%)   
#> 
#> Congenital, familial and genetic disorders
#>   Total number of patients with at least one adverse event                   3 (60%)             1 (20%)             4 (40%)   
#> 
#>   Total number of events                                                        4                   1                   5      
#> 
#>   Congenital aortic dilatation                                                  0                1 (20%)             1 (10%)   
#> 
#>   Congenital cystic disease of liver                                         1 (20%)                0                1 (10%)   
#> 
#>   Craniotabes                                                                1 (20%)                0                1 (10%)   
#> 
#>   Enteric duplication                                                        1 (20%)                0                1 (10%)   
#> 
#>   Tibial agenesis                                                            1 (20%)                0                1 (10%)   
#> 
#> Infections and infestations
#>   Total number of patients with at least one adverse event                   3 (60%)             1 (20%)             4 (40%)   
#> 
#>   Total number of events                                                        5                   1                   6      
#> 
#>   Gallbladder empyema                                                        1 (20%)                0                1 (10%)   
#> 
#>   Mucormycosis                                                               1 (20%)                0                1 (10%)   
#> 
#>   Pleural infection                                                          1 (20%)                0                1 (10%)   
#> 
#>   Retroperitoneal infection                                                  1 (20%)                0                1 (10%)   
#> 
#>   Tuberculosis of genitourinary system                                          0                1 (20%)             1 (10%)   
#> 
#>   Viral rash                                                                 1 (20%)                0                1 (10%)   
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of patients with at least one adverse event                   2 (40%)             2 (40%)             4 (40%)   
#> 
#>   Total number of events                                                        2                   2                   4      
#> 
#>   Harlequin syndrome                                                            0                1 (20%)             1 (10%)   
#> 
#>   Nail pitting                                                               1 (20%)                0                1 (10%)   
#> 
#>   Skin sensitisation                                                            0                1 (20%)             1 (10%)   
#> 
#>   Vaccination site eczema                                                    1 (20%)                0                1 (10%)   
#> 
#> General disorders and administration site conditions
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Catheter site oedema                                                          0                1 (20%)             1 (10%)   
#> 
#>   Membrane stabilising effect                                                1 (20%)                0                1 (10%)   
#> 
#> Immune system disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Dermatomyositis                                                               0                1 (20%)             1 (10%)   
#> 
#>   Myasthenia gravis crisis                                                   1 (20%)                0                1 (10%)   
#> 
#> Injury, poisoning and procedural complications
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Instillation site oedema                                                   1 (20%)                0                1 (10%)   
#> 
#>   Ligament rupture                                                              0                1 (20%)             1 (10%)   
#> 
#> Psychiatric disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Somatic symptom disorder                                                   1 (20%)                0                1 (10%)   
#> 
#>   Tourette's disorder                                                           0                1 (20%)             1 (10%)   
#> 
#> Vascular disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Instillation site thrombosis                                                  0                1 (20%)             1 (10%)   
#> 
#>   Necrotising herpetic retinopathy                                           1 (20%)                0                1 (10%)   
#> 
#> Blood and lymphatic system disorders
#>   Total number of patients with at least one adverse event                      0                1 (20%)             1 (10%)   
#> 
#>   Total number of events                                                        0                   1                   1      
#> 
#>   Blood disorder                                                                0                1 (20%)             1 (10%)   
#> 
#> Cardiac disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Tricuspid valve sclerosis                                                  1 (20%)                0                1 (10%)   
#> 
#> Endocrine disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Emphysematous pyelonephritis                                               1 (20%)                0                1 (10%)   
#> 
#> Gastrointestinal disorders
#>   Total number of patients with at least one adverse event                      0                1 (20%)             1 (10%)   
#> 
#>   Total number of events                                                        0                   1                   1      
#> 
#>   Abscess intestinal                                                            0                1 (20%)             1 (10%)   
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Muscle contusion                                                           1 (20%)                0                1 (10%)   
#> 
#> Renal and urinary disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Alkaptonuria                                                               1 (20%)                0                1 (10%)   
#> 
#> Reproductive system and breast disorders
#>   Total number of patients with at least one adverse event                      0                1 (20%)             1 (10%)   
#> 
#>   Total number of events                                                        0                   1                   1      
#> 
#>   Bleeding anovulatory                                                          0                1 (20%)             1 (10%)   
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of patients with at least one adverse event                      0                1 (20%)             1 (10%)   
#> 
#>   Total number of events                                                        0                   1                   1      
#> 
#>   Hypopharyngeal cancer stage III                                               0                1 (20%)             1 (10%)   
#> 
#> Surgical and medical procedures
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Colporrhaphy                                                               1 (20%)                0                1 (10%)