Create an Adverse Events (AET02) Table

t_ae returns adverse events according to STREAM format AET02

t_ae(class, term, id, col_by, total = "All Patients", sort_by = "count",
  ...)

Arguments

class
system organ class variable.

term
preferred term variable.

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing.

total
character string that will be used as a label for a column with pooled total population, default here is "All Patients", if set to "NONE" then the "All Patients" column is suppressed.

sort_by
character string that defines the ordering of the class and term variables in the output table, options: "alphabetical" or "count", default here is set to "count"

Value

rtable object

Examples

# Simple example
library(tibble)
library(dplyr)
library(rtables)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1,2), c(3,7))])
)

ae_lookup <- tribble(
~CLASS,         ~TERM,   ~GRADE,
"cl A",   "trm A_1/2",        1,
"cl A",   "trm A_2/2",        2,
"cl B",   "trm B_1/3",        2,
"cl B",   "trm B_2/3",        3,
"cl B",   "trm B_3/3",        1,
"cl C",   "trm C_1/1",        1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2,2,2,3,3,4,4,4,4,5,6,6,7,7)]
  ),
  ae_lookup[c(1,1,2,6,4,2,2,3,4,2,1,5,4,6),]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ae(
  class = ANL$CLASS,
  term =  ANL$TERM,
  id = ANL$USUBJID,
  col_by = factor(ANL$ARM),
  total = "All Patients",
)

tbl

#> MedDRA System Organ Class                                            ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                              (N=3)               (N=7)               (N=10)   
#> ----------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event           2 (66.7%)           4 (57.1%)            6 (60%)   
#> 
#> Overall total number of events                                         5                   9                   14     
#> cl A
#>   Total number of patients with at least one adverse event         1 (33.3%)           3 (42.9%)            4 (40%)   
#> 
#>   Total number of events                                               3                   4                   7      
#> 
#>   trm A_2/2                                                        1 (33.3%)           2 (28.6%)            3 (30%)   
#> 
#>   trm A_1/2                                                        1 (33.3%)           1 (14.3%)            2 (20%)   
#> 
#> cl B
#>   Total number of patients with at least one adverse event         1 (33.3%)           3 (42.9%)            4 (40%)   
#> 
#>   Total number of events                                               1                   4                   5      
#> 
#>   trm B_2/3                                                        1 (33.3%)           2 (28.6%)            3 (30%)   
#> 
#>   trm B_1/3                                                            0               1 (14.3%)            1 (10%)   
#> 
#>   trm B_3/3                                                            0               1 (14.3%)            1 (10%)   
#> 
#> cl C
#>   Total number of patients with at least one adverse event         1 (33.3%)           1 (14.3%)            2 (20%)   
#> 
#>   Total number of events                                               1                   1                   2      
#> 
#>   trm C_1/1                                                        1 (33.3%)           1 (14.3%)            2 (20%)   

# Simple example 2

library(dplyr)
library(rtables)
library(random.cdisc.data)

data <- left_join(radam("AAE", N=10),radam("ADSL", N=10))

#> Joining, by = c("STUDYID", "USUBJID")

tbl4 <- t_ae(
  class = data$AEBODSYS,
  term =  data$AEDECOD,
  id = data$USUBJID,
  col_by = factor(data$ARM),
  total = "None",
  sort_by = "alphabetical"
)

tbl4

#> MedDRA System Organ Class                                                    ARM A            ARM B            None   
#>   MedDRA Preferred Term                                                      (N=5)            (N=5)           (N=10)  
#> ----------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event                   5 (100%)         5 (100%)         10 (100%)
#> 
#> Overall total number of events                                                32               33               65    
#> Blood and lymphatic system disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       2                0                2    
#> 
#>   Chronic lymphocytic leukaemia refractory                                  1 (20%)             0             1 (10%) 
#> 
#>   Mastocytic leukaemia                                                      1 (20%)             0             1 (10%) 
#> 
#> Congenital, familial and genetic disorders
#>   Total number of patients with at least one adverse event                  2 (40%)          1 (20%)          3 (30%) 
#> 
#>   Total number of events                                                       3                1                4    
#> 
#>   Benign familial neonatal convulsions                                         0             1 (20%)          1 (10%) 
#> 
#>   Harlequin foetus                                                          1 (20%)             0             1 (10%) 
#> 
#>   Left-to-right cardiac shunt                                               1 (20%)             0             1 (10%) 
#> 
#>   Oculodentodigital dysplasia                                               1 (20%)             0             1 (10%) 
#> 
#> Endocrine disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Testicular failure primary                                                1 (20%)             0             1 (10%) 
#> 
#> Gastrointestinal disorders
#>   Total number of patients with at least one adverse event                  4 (80%)          2 (40%)          6 (60%) 
#> 
#>   Total number of events                                                       5                2                7    
#> 
#>   Faecalith                                                                    0             1 (20%)          1 (10%) 
#> 
#>   Gastroenteritis rotavirus                                                 1 (20%)             0             1 (10%) 
#> 
#>   Gastrointestinal tract adenoma                                            1 (20%)             0             1 (10%) 
#> 
#>   Haemorrhagic ascites                                                      1 (20%)             0             1 (10%) 
#> 
#>   Peritoneal candidiasis                                                       0             1 (20%)          1 (10%) 
#> 
#>   Peritoneal chlamydia infection                                            1 (20%)             0             1 (10%) 
#> 
#>   Rectocele                                                                 1 (20%)             0             1 (10%) 
#> 
#> General disorders and administration site conditions
#>   Total number of patients with at least one adverse event                     0             2 (40%)          2 (20%) 
#> 
#>   Total number of events                                                       0                4                4    
#> 
#>   Cardiac complication associated with device                                  0             1 (20%)          1 (10%) 
#> 
#>   Fistula inflammation                                                         0             1 (20%)          1 (10%) 
#> 
#>   Injection site cyst                                                          0             1 (20%)          1 (10%) 
#> 
#>   Polyp                                                                        0             1 (20%)          1 (10%) 
#> 
#> Hepatobiliary disorders
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       1                0                1    
#> 
#>   Gallbladder cancer stage 0                                                1 (20%)             0             1 (10%) 
#> 
#> Infections and infestations
#>   Total number of patients with at least one adverse event                  2 (40%)          2 (40%)          4 (40%) 
#> 
#>   Total number of events                                                       2                3                5    
#> 
#>   Bone abscess                                                                 0             1 (20%)          1 (10%) 
#> 
#>   Bordetella infection                                                      1 (20%)             0             1 (10%) 
#> 
#>   Cytomegalovirus syndrome                                                     0             1 (20%)          1 (10%) 
#> 
#>   Nipple infection                                                          1 (20%)             0             1 (10%) 
#> 
#>   Viral hepatitis carrier                                                      0             1 (20%)          1 (10%) 
#> 
#> Injury, poisoning and procedural complications
#>   Total number of patients with at least one adverse event                     0             4 (80%)          4 (40%) 
#> 
#>   Total number of events                                                       0                5                5    
#> 
#>   Chemical burn of gastrointestinal tract                                      0             1 (20%)          1 (10%) 
#> 
#>   Febrile nonhaemolytic transfusion reaction                                   0             1 (20%)          1 (10%) 
#> 
#>   Hyphaema                                                                     0             1 (20%)          1 (10%) 
#> 
#>   Occupational exposure to toxic agent                                         0             1 (20%)          1 (10%) 
#> 
#>   Systemic leakage                                                             0             1 (20%)          1 (10%) 
#> 
#> Investigations
#>   Total number of patients with at least one adverse event                  3 (60%)          3 (60%)          6 (60%) 
#> 
#>   Total number of events                                                       4                5                9    
#> 
#>   Biopsy parathyroid gland normal                                              0             1 (20%)          1 (10%) 
#> 
#>   Blood parathyroid hormone abnormal                                        1 (20%)             0             1 (10%) 
#> 
#>   Empedobacter test positive                                                1 (20%)             0             1 (10%) 
#> 
#>   Endothelin decreased                                                      1 (20%)             0             1 (10%) 
#> 
#>   Hepatitis C RNA decreased                                                    0             1 (20%)          1 (10%) 
#> 
#>   Hydroxyproline decreased                                                     0             1 (20%)          1 (10%) 
#> 
#>   Influenza A virus test                                                    1 (20%)             0             1 (10%) 
#> 
#>   Light chain analysis decreased                                               0             1 (20%)          1 (10%) 
#> 
#>   Urine viscosity abnormal                                                     0             1 (20%)          1 (10%) 
#> 
#> Metabolism and nutrition disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Pantothenic acid deficiency                                                  0             1 (20%)          1 (10%) 
#> 
#>   Type III hyperlipidaemia                                                  1 (20%)             0             1 (10%) 
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of patients with at least one adverse event                     0             1 (20%)          1 (10%) 
#> 
#>   Total number of events                                                       0                1                1    
#> 
#>   Schinzel-Giedion syndrome                                                    0             1 (20%)          1 (10%) 
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of patients with at least one adverse event                  1 (20%)             0             1 (10%) 
#> 
#>   Total number of events                                                       2                0                2    
#> 
#>   Multicentric reticulohistiocytosis                                        1 (20%)             0             1 (10%) 
#> 
#>   Oesophageal squamous cell carcinoma metastatic                            1 (20%)             0             1 (10%) 
#> 
#> Nervous system disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                2                3    
#> 
#>   Application site hypoaesthesia                                            1 (20%)             0             1 (10%) 
#> 
#>   Exertional headache                                                          0             1 (20%)          1 (10%) 
#> 
#>   Ganglioneuroblastoma                                                         0             1 (20%)          1 (10%) 
#> 
#> Renal and urinary disorders
#>   Total number of patients with at least one adverse event                     0             2 (40%)          2 (20%) 
#> 
#>   Total number of events                                                       0                2                2    
#> 
#>   Calyceal diverticulum                                                        0             1 (20%)          1 (10%) 
#> 
#>   Pulmonary renal syndrome                                                     0             1 (20%)          1 (10%) 
#> 
#> Reproductive system and breast disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       1                1                2    
#> 
#>   Prostatic abscess                                                         1 (20%)             0             1 (10%) 
#> 
#>   Radiation prostatitis                                                        0             1 (20%)          1 (10%) 
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          2 (40%)          3 (30%) 
#> 
#>   Total number of events                                                       1                2                3    
#> 
#>   Non-small cell lung cancer stage IIIA                                     1 (20%)             0             1 (10%) 
#> 
#>   Procedural pneumothorax                                                      0             1 (20%)          1 (10%) 
#> 
#>   Vocal cord leukoplakia                                                       0             1 (20%)          1 (10%) 
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of patients with at least one adverse event                  2 (40%)             0             2 (20%) 
#> 
#>   Total number of events                                                       4                0                4    
#> 
#>   Eczema herpeticum                                                         1 (20%)             0             1 (10%) 
#> 
#>   Hyper IgD syndrome                                                        1 (20%)             0             1 (10%) 
#> 
#>   Medical device site erosion                                               1 (20%)             0             1 (10%) 
#> 
#>   Skin swelling                                                             1 (20%)             0             1 (10%) 
#> 
#> Surgical and medical procedures
#>   Total number of patients with at least one adverse event                  1 (20%)          2 (40%)          3 (30%) 
#> 
#>   Total number of events                                                       2                3                5    
#> 
#>   Cholecystostomy                                                           1 (20%)             0             1 (10%) 
#> 
#>   Isotope therapy to pleura                                                    0             1 (20%)          1 (10%) 
#> 
#>   Lymphadenectomy                                                              0             1 (20%)          1 (10%) 
#> 
#>   Repair of diaphragm injury                                                1 (20%)             0             1 (10%) 
#> 
#>   Vagal nerve stimulator removal                                               0             1 (20%)          1 (10%) 
#> 
#> Vascular disorders
#>   Total number of patients with at least one adverse event                  1 (20%)          1 (20%)          2 (20%) 
#> 
#>   Total number of events                                                       2                1                3    
#> 
#>   Catheter site phlebitis                                                   1 (20%)             0             1 (10%) 
#> 
#>   Hepatic ischaemia                                                            0             1 (20%)          1 (10%) 
#> 
#>   Ocular vascular disorder                                                  1 (20%)             0             1 (10%) 

data <- read_bce("/opt/BIOSTAT/home_ext2/qit3/cdt70194/go39733/libraries/adae.sas7bdat")

#> Error in read_bce("/opt/BIOSTAT/home_ext2/qit3/cdt70194/go39733/libraries/adae.sas7bdat"): could not find function "read_bce"

tbl5 <- t_ae(
  class = data$AEBODSYS,
  term =  data$AEDECOD,
  id = data$USUBJID,
  col_by = factor(data$ARM)
)

tbl5

#> MedDRA System Organ Class                                                     ARM A               ARM B            All Patients
#>   MedDRA Preferred Term                                                       (N=5)               (N=5)               (N=10)   
#> -------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one adverse event                     5 (100%)            5 (100%)           10 (100%)  
#> 
#> Overall total number of events                                                  32                  33                  65     
#> Gastrointestinal disorders
#>   Total number of patients with at least one adverse event                   4 (80%)             2 (40%)             6 (60%)   
#> 
#>   Total number of events                                                        5                   2                   7      
#> 
#>   Faecalith                                                                     0                1 (20%)             1 (10%)   
#> 
#>   Gastroenteritis rotavirus                                                  1 (20%)                0                1 (10%)   
#> 
#>   Gastrointestinal tract adenoma                                             1 (20%)                0                1 (10%)   
#> 
#>   Haemorrhagic ascites                                                       1 (20%)                0                1 (10%)   
#> 
#>   Peritoneal candidiasis                                                        0                1 (20%)             1 (10%)   
#> 
#>   Peritoneal chlamydia infection                                             1 (20%)                0                1 (10%)   
#> 
#>   Rectocele                                                                  1 (20%)                0                1 (10%)   
#> 
#> Investigations
#>   Total number of patients with at least one adverse event                   3 (60%)             3 (60%)             6 (60%)   
#> 
#>   Total number of events                                                        4                   5                   9      
#> 
#>   Biopsy parathyroid gland normal                                               0                1 (20%)             1 (10%)   
#> 
#>   Blood parathyroid hormone abnormal                                         1 (20%)                0                1 (10%)   
#> 
#>   Empedobacter test positive                                                 1 (20%)                0                1 (10%)   
#> 
#>   Endothelin decreased                                                       1 (20%)                0                1 (10%)   
#> 
#>   Hepatitis C RNA decreased                                                     0                1 (20%)             1 (10%)   
#> 
#>   Hydroxyproline decreased                                                      0                1 (20%)             1 (10%)   
#> 
#>   Influenza A virus test                                                     1 (20%)                0                1 (10%)   
#> 
#>   Light chain analysis decreased                                                0                1 (20%)             1 (10%)   
#> 
#>   Urine viscosity abnormal                                                      0                1 (20%)             1 (10%)   
#> 
#> Infections and infestations
#>   Total number of patients with at least one adverse event                   2 (40%)             2 (40%)             4 (40%)   
#> 
#>   Total number of events                                                        2                   3                   5      
#> 
#>   Bone abscess                                                                  0                1 (20%)             1 (10%)   
#> 
#>   Bordetella infection                                                       1 (20%)                0                1 (10%)   
#> 
#>   Cytomegalovirus syndrome                                                      0                1 (20%)             1 (10%)   
#> 
#>   Nipple infection                                                           1 (20%)                0                1 (10%)   
#> 
#>   Viral hepatitis carrier                                                       0                1 (20%)             1 (10%)   
#> 
#> Injury, poisoning and procedural complications
#>   Total number of patients with at least one adverse event                      0                4 (80%)             4 (40%)   
#> 
#>   Total number of events                                                        0                   5                   5      
#> 
#>   Chemical burn of gastrointestinal tract                                       0                1 (20%)             1 (10%)   
#> 
#>   Febrile nonhaemolytic transfusion reaction                                    0                1 (20%)             1 (10%)   
#> 
#>   Hyphaema                                                                      0                1 (20%)             1 (10%)   
#> 
#>   Occupational exposure to toxic agent                                          0                1 (20%)             1 (10%)   
#> 
#>   Systemic leakage                                                              0                1 (20%)             1 (10%)   
#> 
#> Congenital, familial and genetic disorders
#>   Total number of patients with at least one adverse event                   2 (40%)             1 (20%)             3 (30%)   
#> 
#>   Total number of events                                                        3                   1                   4      
#> 
#>   Benign familial neonatal convulsions                                          0                1 (20%)             1 (10%)   
#> 
#>   Harlequin foetus                                                           1 (20%)                0                1 (10%)   
#> 
#>   Left-to-right cardiac shunt                                                1 (20%)                0                1 (10%)   
#> 
#>   Oculodentodigital dysplasia                                                1 (20%)                0                1 (10%)   
#> 
#> Respiratory, thoracic and mediastinal disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             2 (40%)             3 (30%)   
#> 
#>   Total number of events                                                        1                   2                   3      
#> 
#>   Non-small cell lung cancer stage IIIA                                      1 (20%)                0                1 (10%)   
#> 
#>   Procedural pneumothorax                                                       0                1 (20%)             1 (10%)   
#> 
#>   Vocal cord leukoplakia                                                        0                1 (20%)             1 (10%)   
#> 
#> Surgical and medical procedures
#>   Total number of patients with at least one adverse event                   1 (20%)             2 (40%)             3 (30%)   
#> 
#>   Total number of events                                                        2                   3                   5      
#> 
#>   Cholecystostomy                                                            1 (20%)                0                1 (10%)   
#> 
#>   Isotope therapy to pleura                                                     0                1 (20%)             1 (10%)   
#> 
#>   Lymphadenectomy                                                               0                1 (20%)             1 (10%)   
#> 
#>   Repair of diaphragm injury                                                 1 (20%)                0                1 (10%)   
#> 
#>   Vagal nerve stimulator removal                                                0                1 (20%)             1 (10%)   
#> 
#> General disorders and administration site conditions
#>   Total number of patients with at least one adverse event                      0                2 (40%)             2 (20%)   
#> 
#>   Total number of events                                                        0                   4                   4      
#> 
#>   Cardiac complication associated with device                                   0                1 (20%)             1 (10%)   
#> 
#>   Fistula inflammation                                                          0                1 (20%)             1 (10%)   
#> 
#>   Injection site cyst                                                           0                1 (20%)             1 (10%)   
#> 
#>   Polyp                                                                         0                1 (20%)             1 (10%)   
#> 
#> Metabolism and nutrition disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Pantothenic acid deficiency                                                   0                1 (20%)             1 (10%)   
#> 
#>   Type III hyperlipidaemia                                                   1 (20%)                0                1 (10%)   
#> 
#> Nervous system disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   2                   3      
#> 
#>   Application site hypoaesthesia                                             1 (20%)                0                1 (10%)   
#> 
#>   Exertional headache                                                           0                1 (20%)             1 (10%)   
#> 
#>   Ganglioneuroblastoma                                                          0                1 (20%)             1 (10%)   
#> 
#> Renal and urinary disorders
#>   Total number of patients with at least one adverse event                      0                2 (40%)             2 (20%)   
#> 
#>   Total number of events                                                        0                   2                   2      
#> 
#>   Calyceal diverticulum                                                         0                1 (20%)             1 (10%)   
#> 
#>   Pulmonary renal syndrome                                                      0                1 (20%)             1 (10%)   
#> 
#> Reproductive system and breast disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        1                   1                   2      
#> 
#>   Prostatic abscess                                                          1 (20%)                0                1 (10%)   
#> 
#>   Radiation prostatitis                                                         0                1 (20%)             1 (10%)   
#> 
#> Skin and subcutaneous tissue disorders
#>   Total number of patients with at least one adverse event                   2 (40%)                0                2 (20%)   
#> 
#>   Total number of events                                                        4                   0                   4      
#> 
#>   Eczema herpeticum                                                          1 (20%)                0                1 (10%)   
#> 
#>   Hyper IgD syndrome                                                         1 (20%)                0                1 (10%)   
#> 
#>   Medical device site erosion                                                1 (20%)                0                1 (10%)   
#> 
#>   Skin swelling                                                              1 (20%)                0                1 (10%)   
#> 
#> Vascular disorders
#>   Total number of patients with at least one adverse event                   1 (20%)             1 (20%)             2 (20%)   
#> 
#>   Total number of events                                                        2                   1                   3      
#> 
#>   Catheter site phlebitis                                                    1 (20%)                0                1 (10%)   
#> 
#>   Hepatic ischaemia                                                             0                1 (20%)             1 (10%)   
#> 
#>   Ocular vascular disorder                                                   1 (20%)                0                1 (10%)   
#> 
#> Blood and lymphatic system disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        2                   0                   2      
#> 
#>   Chronic lymphocytic leukaemia refractory                                   1 (20%)                0                1 (10%)   
#> 
#>   Mastocytic leukaemia                                                       1 (20%)                0                1 (10%)   
#> 
#> Endocrine disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Testicular failure primary                                                 1 (20%)                0                1 (10%)   
#> 
#> Hepatobiliary disorders
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        1                   0                   1      
#> 
#>   Gallbladder cancer stage 0                                                 1 (20%)                0                1 (10%)   
#> 
#> Musculoskeletal and connective tissue disorders
#>   Total number of patients with at least one adverse event                      0                1 (20%)             1 (10%)   
#> 
#>   Total number of events                                                        0                   1                   1      
#> 
#>   Schinzel-Giedion syndrome                                                     0                1 (20%)             1 (10%)   
#> 
#> Neoplasms benign, malignant and unspecified (incl cysts and polyps)
#>   Total number of patients with at least one adverse event                   1 (20%)                0                1 (10%)   
#> 
#>   Total number of events                                                        2                   0                   2      
#> 
#>   Multicentric reticulohistiocytosis                                         1 (20%)                0                1 (10%)   
#> 
#>   Oesophageal squamous cell carcinoma metastatic                             1 (20%)                0                1 (10%)