Adverse Events Table by Highest NCI CTCAE Grade

t_ae_ctc_v2 returns adverse events sorted by highest NCI (National Cancer Institute) CTCAE (common terminology criteria for adverse events) grade. It corresponds to STREAM template AET01.

t_ae_ctc_v2(
  class,
  term,
  id,
  grade,
  col_by,
  total = "All Patients",
  grade_levels = as.factor(1:5)
)

Arguments

class
system organ class variable.

term
preferred term variable.

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

grade
grade of adverse event. For factors, it is assumed that intensity corresponds to the order of the factor levels. If that is not the case, see grade_levels. For character or numeric, grade_levels is required.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing. See 'Examples'.

total
character string that will be used as a label for a column with pooled total population, default is "All Patients", if set to NULL then the "All Patients" column is suppressed.

grade_levels
a factor. The values of the factor define the ordering of the rows in the resulting table. The levels of the factor define the severity of the grade. For example, factor(c("c", "b", "a"), levels = c("a", "b", "c")) will display the most severe grade "c" at the top of the table, the least severe grade "a" at the bottom. If grade is a factor, grade_levels will overwrite the level orders in grade. Default is as.factor(1:5). If set to NULL, it is assumed that intensity corresponds to the order of the factor levels of grade. If grade is not a factor, grade_levels is required.

Details

t_ae_ctc_v2 counts patients according to adverse events (AEs) of greatest intensity for system organ class (SOC) and overall rows and includes percentages based on the total number of patients in the column heading (i.e. "N=nnn"). If the intention is to use patients number from subject level dataset as N for percentage calculation then adverse events dataset should be left joined to subject level dataset and the col_by variable should be dropped from adverse events dataset, see the example. Otherwise, N will be derived using adverse events dataset. At the preferred term (PT) level, multiple events within a patient of the same PT are counted once using the greatest intensity reported.

t_ae_ctc_v2 removes any non-complete records, e.g. if class or term are missing. If the intent is to preserve such records, then impute missing values before using t_ae_ctc_v2.

t_ae_ctc_v2 orders data by "All Patients" column from the most commonly reported SOC to the least frequent one. Within SOC, it sorts by decreasing frequency of PT. It brakes ties using SOC/PT names in alphabetical order.

t_ae_ctc_v2 fills in col_by and grade with 0 value in case there was no AEs reported for particular col_by and/or grade category. Use grade_levels to modify the range of existing grades. If data does not have any records with grade 5 and the intent is to show only grades 1-4 rows then use grade_levels = as.factor(1:4).

this is an equivalent of the STREAM output %stream_t_summary(templates = aet04) (http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet04.html)

Examples

# Simple example
library(dplyr)

ASL <- tibble(
  USUBJID = paste0("id-", 1:10),
  ARM = paste("ARM", LETTERS[rep(c(1, 2), c(3, 7))])
)

ae_lookup <- tribble(
  ~CLASS, ~TERM, ~GRADE,
  "cl A", "trm A_1/2", 1,
  "cl A", "trm A_2/2", 2,
  "cl B", "trm B_1/3", 2,
  "cl B", "trm B_2/3", 3,
  "cl B", "trm B_3/3", 1,
  "cl C", "trm C_1/1", 1
)

AAE <- cbind(
  tibble(
    USUBJID = ASL$USUBJID[c(2, 2, 2, 3, 3, 4, 4, 4, 4, 5, 6, 6, 7, 7)]
  ),
  ae_lookup[c(1, 1, 2, 6, 4, 2, 2, 3, 4, 2, 1, 5, 4, 6), ]
)

ANL <- left_join(ASL, AAE, by = "USUBJID")

tbl <- t_ae_ctc_v2(
  class = ANL$CLASS,
  term = ANL$TERM,
  id = ANL$USUBJID,
  grade = ANL$GRADE,
  col_by = factor(ANL$ARM),
  total = "All Patients",
  grade_levels = as.factor(1:3)
)
tbl

#> ANL$CLASS                                                    ARM A                    ARM B                All Patients   
#>   ANL$TERM                        ANL$GRADE                  (N=3)                    (N=7)                   (N=10)      
#> --------------------------------------------------------------------------------------------------------------------------
#> - Any adverse events -
#>   - Overall -                 - - Any Grade - -            2 (66.7%)                4 (57.1%)                 6 (60%)     
#>                                       1                        0                    1 (14.3%)                 1 (10%)     
#>                                       2                    1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       3                    1 (33.3%)                2 (28.6%)                 3 (30%)     
#> 
#> cl A
#>   - Overall -                 - - Any Grade - -            1 (33.3%)                3 (42.9%)                 4 (40%)     
#>                                       1                        0                    1 (14.3%)                 1 (10%)     
#>                                       2                    1 (33.3%)                2 (28.6%)                 3 (30%)     
#>                                       3                        0                        0                        0        
#> 
#>   trm A_2/2                   - - Any Grade - -            1 (33.3%)                2 (28.6%)                 3 (30%)     
#>                                       1                        0                        0                        0        
#>                                       2                    1 (33.3%)                2 (28.6%)                 3 (30%)     
#>                                       3                        0                        0                        0        
#> 
#>   trm A_1/2                   - - Any Grade - -            1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       1                    1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       2                        0                        0                        0        
#>                                       3                        0                        0                        0        
#> 
#> cl B
#>   - Overall -                 - - Any Grade - -            1 (33.3%)                3 (42.9%)                 4 (40%)     
#>                                       1                        0                    1 (14.3%)                 1 (10%)     
#>                                       2                        0                        0                        0        
#>                                       3                    1 (33.3%)                2 (28.6%)                 3 (30%)     
#> 
#>   trm B_2/3                   - - Any Grade - -            1 (33.3%)                2 (28.6%)                 3 (30%)     
#>                                       1                        0                        0                        0        
#>                                       2                        0                        0                        0        
#>                                       3                    1 (33.3%)                2 (28.6%)                 3 (30%)     
#> 
#>   trm B_1/3                   - - Any Grade - -                0                    1 (14.3%)                 1 (10%)     
#>                                       1                        0                        0                        0        
#>                                       2                        0                    1 (14.3%)                 1 (10%)     
#>                                       3                        0                        0                        0        
#> 
#>   trm B_3/3                   - - Any Grade - -                0                    1 (14.3%)                 1 (10%)     
#>                                       1                        0                    1 (14.3%)                 1 (10%)     
#>                                       2                        0                        0                        0        
#>                                       3                        0                        0                        0        
#> 
#> cl C
#>   - Overall -                 - - Any Grade - -            1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       1                    1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       2                        0                        0                        0        
#>                                       3                        0                        0                        0        
#> 
#>   trm C_1/1                   - - Any Grade - -            1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       1                    1 (33.3%)                1 (14.3%)                 2 (20%)     
#>                                       2                        0                        0                        0        
#>                                       3                        0                        0                        0        

library(random.cdisc.data)
library(dplyr)

ASL <- radsl(N = 10, seed = 1)
AAE <- radae(ASL, seed = 1)

tbl <- t_ae_ctc_v2(
    class = AAE$AEBODSYS,
    term = AAE$AEDECOD,
    id = AAE$USUBJID,
    grade = AAE$AETOXGR,
    col_by = factor(AAE$ARM),
    total = "All Patients",
    grade_levels = as.factor(1:5)
  )
tbl

#> Body System or Organ Class                                              A: Drug X                     B: Placebo                   C: Combination                  All Patients      
#>   Dictionary-Derived Term         Analysis Toxicity Grade                 (N=6)                          (N=2)                          (N=2)                         (N=10)         
#> -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> - Any adverse events -
#>   - Overall -                        - - Any Grade - -                  6 (100%)                       2 (100%)                       2 (100%)                       10 (100%)       
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                          2 (33.3%)                          0                              0                           2 (20%)        
#>                                              4                          1 (16.7%)                          0                           1 (50%)                        2 (20%)        
#>                                              5                           3 (50%)                       2 (100%)                        1 (50%)                        6 (60%)        
#> 
#> 4
#>   - Overall -                        - - Any Grade - -                  4 (66.7%)                       1 (50%)                       2 (100%)                        7 (70%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                          4 (66.7%)                       1 (50%)                       2 (100%)                        7 (70%)        
#>                                              5                              0                              0                              0                              0           
#> 
#>   6                                  - - Any Grade - -                  4 (66.7%)                       1 (50%)                       2 (100%)                        7 (70%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                          4 (66.7%)                       1 (50%)                       2 (100%)                        7 (70%)        
#>                                              5                              0                              0                              0                              0           
#> 
#> 3
#>   - Overall -                        - - Any Grade - -                   3 (50%)                        1 (50%)                       2 (100%)                        6 (60%)        
#>                                              1                          2 (33.3%)                          0                              0                           2 (20%)        
#>                                              2                              0                              0                              0                              0           
#>                                              3                          1 (16.7%)                       1 (50%)                       2 (100%)                        4 (40%)        
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   4                                  - - Any Grade - -                  1 (16.7%)                       1 (50%)                       2 (100%)                        4 (40%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                          1 (16.7%)                       1 (50%)                       2 (100%)                        4 (40%)        
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   5                                  - - Any Grade - -                  2 (33.3%)                          0                              0                           2 (20%)        
#>                                              1                          2 (33.3%)                          0                              0                           2 (20%)        
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#> 6
#>   - Overall -                        - - Any Grade - -                  4 (66.7%)                          0                          2 (100%)                        6 (60%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                          2 (33.3%)                          0                           1 (50%)                        3 (30%)        
#>                                              4                              0                              0                              0                              0           
#>                                              5                          2 (33.3%)                          0                           1 (50%)                        3 (30%)        
#> 
#>   9                                  - - Any Grade - -                  2 (33.3%)                          0                          2 (100%)                        4 (40%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                          2 (33.3%)                          0                          2 (100%)                        4 (40%)        
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   8                                  - - Any Grade - -                  2 (33.3%)                          0                           1 (50%)                        3 (30%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                          2 (33.3%)                          0                           1 (50%)                        3 (30%)        
#> 
#> 1
#>   - Overall -                        - - Any Grade - -                  2 (33.3%)                      2 (100%)                        1 (50%)                        5 (50%)        
#>                                              1                          1 (16.7%)                          0                           1 (50%)                        2 (20%)        
#>                                              2                          1 (16.7%)                      2 (100%)                           0                           3 (30%)        
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   1                                  - - Any Grade - -                  2 (33.3%)                      2 (100%)                        1 (50%)                        5 (50%)        
#>                                              1                          2 (33.3%)                      2 (100%)                        1 (50%)                        5 (50%)        
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   2                                  - - Any Grade - -                  1 (16.7%)                      2 (100%)                           0                           3 (30%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                          1 (16.7%)                      2 (100%)                           0                           3 (30%)        
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#> 5
#>   - Overall -                        - - Any Grade - -                  2 (33.3%)                       1 (50%)                       2 (100%)                        5 (50%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                          2 (33.3%)                       1 (50%)                       2 (100%)                        5 (50%)        
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   7                                  - - Any Grade - -                  2 (33.3%)                       1 (50%)                       2 (100%)                        5 (50%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                          2 (33.3%)                       1 (50%)                       2 (100%)                        5 (50%)        
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#> 7
#>   - Overall -                        - - Any Grade - -                  4 (66.7%)                          0                           1 (50%)                        5 (50%)        
#>                                              1                          4 (66.7%)                          0                           1 (50%)                        5 (50%)        
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#>   10                                 - - Any Grade - -                  4 (66.7%)                          0                           1 (50%)                        5 (50%)        
#>                                              1                          4 (66.7%)                          0                           1 (50%)                        5 (50%)        
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                              0                              0                              0                              0           
#> 
#> 2
#>   - Overall -                        - - Any Grade - -                  2 (33.3%)                      2 (100%)                           0                           4 (40%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                          2 (33.3%)                      2 (100%)                           0                           4 (40%)        
#> 
#>   3                                  - - Any Grade - -                  2 (33.3%)                      2 (100%)                           0                           4 (40%)        
#>                                              1                              0                              0                              0                              0           
#>                                              2                              0                              0                              0                              0           
#>                                              3                              0                              0                              0                              0           
#>                                              4                              0                              0                              0                              0           
#>                                              5                          2 (33.3%)                      2 (100%)                           0                           4 (40%)