AE Overview Summary Table

t_ae_oview returns a summary table of overall AE profile: adverse events, deaths, and withdrawals by trial treatment. It corresponds to STREAM template AET01.

t_ae_oview(
  id,
  class,
  term,
  flags,
  display_id = c("fatal", "ser", "serwd", "serdsm", "relser", "wd", "dsm", "rel",
    "relwd", "reldsm", "ctc35"),
  extra_flag = NULL,
  col_by,
  total = "All Patients"
)

Arguments

id
unique subject identifier variable. If a particular subject has no adverse event then the subject id should be listed where class and term should be set to missing (i.e. NA).

class
system organ class variable.

term
preferred term variable.

flags
dataframe containing flag variables of all events of interest, default variables should include dthfl, dcsreas, aesdth, aeser, aeacn, aerel, aetoxgr to match STREAM output (lowercase column names). Flags dataframe must include dthfl and dcsreas at minimal.

display_id
vector of strings for which analyses to display possible values are: fatal ser serwd serdsm relser wd dsm rel relwd reldsm ctc35

extra_flag
dataframe with flag variables to be tabulated. Only records where values equals to "Y" are counted. Variable names of the dataframe will be used as the summary table row name.

col_by
group variable that will be used for a column header. col_by has to be a factor and can not be missing.

total
character string that will be used as a label for a column with pooled total population, default here is "All Patients", if set to NULL then the "All Patients" column is suppressed.

Value

rtable object

Details

this is an equivalent of the STREAM output %stream_t_summary(templates = aet01) (http://bioportal.roche.com/stream_doc/2_05/um/report_outputs_aet01.html)

The flags dataframe should be taken from the same dataset and in the same order as id, class and term. It should only contain variables needed for the specified flags. All other desired variables should be passed to the extra_flag argument, along with the desired label names (as the variable name of dataframe). Each variable in extra_flag will be tabulated to one row in the summary table, where only "Y" values are counted towards the total.

Examples

library(dplyr)

ANL <- rADAE
flag <- data.frame(
  dthfl = ANL$DTHFL,
  dcsreas = ANL$DCSREAS,
  aesdth = ANL$AESDTH,
  aeser = ANL$AESER,
  aeacn = ANL$AEACN,
  aerel = ANL$AEREL,
  aetoxgr = ANL$AETOXGR
)

extra <- data.frame(
  fatal2 = flag$aesdth,
  ser2 = flag$aeser
)

tbl <- t_ae_oview(
  id = ANL$USUBJID,
  class = ANL$AESOC,
  term = ANL$AEDECOD,
  flags = flag,
  display_id = c(
    "fatal", "ser", "serwd", "serdsm", "relser",
    "wd", "dsm", "rel", "relwd", "reldsm"
  ),
  extra_flag = extra,
  col_by = factor(ANL$ARM),
  total = "All Patients"
)

tbl

#>                                                                     A: Placebo            B: Drug X           C: Combination         All Patients 
#>                                                                       (N=11)                (N=21)                (N=18)                (N=50)    
#> --------------------------------------------------------------------------------------------------------------------------------------------------
#> Total number of patients with at least one AE                       11 (100%)             21 (100%)             18 (100%)             50 (100%)   
#> Total number of AEs                                                     65                   133                    99                   297      
#> Total number of deaths                                              2 (18.18%)            4 (19.05%)            1 (5.56%)              7 (14%)    
#> Total number of patients withdrawn from study due to an AE              0                 1 (4.76%)                 0                   1 (2%)    
#> 
#> Total number of patients with at least one
#>   AE with fatal outcome                                                 0                 1 (4.76%)                 0                   1 (2%)    
#>   Serious AE                                                        8 (72.73%)           16 (76.19%)           12 (66.67%)             36 (72%)   
#>   Serious AE leading to withdrawal from treatment                       0                 2 (9.52%)                 0                   2 (4%)    
#>   Serious AE leading to dose modification/interruption              2 (18.18%)            5 (23.81%)            5 (27.78%)             12 (24%)   
#>   Related Serious AE                                                1 (9.09%)             4 (19.05%)            3 (16.67%)             8 (16%)    
#>   AE leading to withdrawal from treatment                               0                 2 (9.52%)                 0                   2 (4%)    
#>   AE leading to dose modification/interruption                      9 (81.82%)           17 (80.95%)           13 (72.22%)             39 (78%)   
#>   Related AE                                                        7 (63.64%)            13 (61.9%)             9 (50%)               29 (58%)   
#>   Related AE leading to withdrawal from treatment                       0                     0                     0                     0       
#>   Related AE leading to dose modification/interruption              3 (27.27%)            5 (23.81%)            2 (11.11%)             10 (20%)   
#>   fatal2                                                                0                 1 (4.76%)                 0                   1 (2%)    
#>   ser2                                                              8 (72.73%)           16 (76.19%)           12 (66.67%)             36 (72%)